tests: move shared helper methods to \Helper\Helper
