+++ tests/_support/Helper/Functional.php
@@ -6,16 +6,4 @@ namespace Helper;
 
 class Functional extends \Codeception\Module {
 
-	public function reset_utils_properties() {
-		$reflection = new ReflectionClass( 'WPGraphQL\Login\Utils\Utils' );
-		// Reset Providers.
-		$property = $reflection->getProperty( 'providers' );
-		$property->setAccessible( true );
-		$property->setValue( [] );
-
-		// Reset Settings.
-		$property = $reflection->getProperty( 'settings' );
-		$property->setAccessible( true );
-		$property->setValue( [] );
-	}
 }
