+++ tests/functional.suite.dist.yml
@@ -7,6 +7,7 @@ actor: FunctionalTester
 modules:
     enabled:
       - \Helper\Functional
+      - \Helper\Helper
       - Asserts
       - REST
       - WPDb
