+++ tests/wpunit.suite.dist.yml
@@ -1,10 +1,11 @@
 # Codeception Test Suite Configuration
 #
 # Suite for unit or integration tests that require WordPress functions and classes.
-actor: WpunitTesterActions
+actor: WpunitTester
 modules:
   enabled:
     - WPLoader
+    - \Helper\Helper
     - \Helper\Wpunit
   disabled:
     - WPDb
